moved pooled donation modal to separate component
